+++ Public/VectorStores/Files/Add-VectorStoreFile.ps1
@@ -24,6 +24,21 @@ function Add-VectorStoreFile {
         [PSTypeName('PSOpenAI.File')]$File,
 
         [Parameter()]
+        [ValidateSet('auto', 'static')]
+        [Alias('chunking_strategy')]
+        [string]$ChunkingStrategy,
+
+        [Parameter()]
+        [ValidateRange(100, 4096)]
+        [Alias('max_chunk_size_tokens')]
+        [int]$MaxChunkSizeTokens = 800,
+
+        [Parameter()]
+        [ValidateRange(0, 4096)]
+        [Alias('chunk_overlap_tokens')]
+        [int]$ChunkOverlapTokens = 400,
+
+        [Parameter()]
         [switch]$PassThru,
 
         [Parameter()]
@@ -94,6 +109,21 @@ function Add-VectorStoreFile {
 
         $PostBody = [System.Collections.Specialized.OrderedDictionary]::new()
         $PostBody.file_id = $FileId
+
+        if ($PSBoundParameters.ContainsKey('ChunkingStrategy')) {
+            $PostBody.chunking_strategy = @{type = $ChunkingStrategy }
+            if ($ChunkingStrategy -eq 'static') {
+                # validation (the overlap must not exceed half of max)
+                if (2 * $ChunkOverlapTokens -gt $MaxChunkSizeTokens) {
+                    Write-Error -Exception ([System.ArgumentException]::new('ChunkOverlapTokens must not exceed half of MaxChunkSizeTokens.'))
+                    return
+                }
+                $PostBody.chunking_strategy.static = @{
+                    max_chunk_size_tokens = $MaxChunkSizeTokens
+                    chunk_overlap_tokens  = $ChunkOverlapTokens
+                }
+            }
+        }
         #endregion
 
         #region Send API Request